fix imports for moved courses feature
