+++ frontend/src/app/features/memorize/courses/workflow-list.component.scss
