--- frontend/src/app/features/workflows/browse/frontend/src/app/features/workflows/browse/browse-workflows.component.scss
@@ -1,477 +1,0 @@
-// frontend/src/app/features/workflows/browse/browse-workflows.component.scss
-
-.browse-workflows {
-  padding: 2rem;
-  max-width: 1400px;
-  margin: 0 auto;
-}
-
-// Header
-.page-header {
-  text-align: center;
-  margin-bottom: 3rem;
-
-  h1 {
-    font-size: 2.5rem;
-    font-weight: 700;
-    color: #1f2937;
-    margin-bottom: 0.5rem;
-  }
-
-  p {
-    font-size: 1.125rem;
-    color: #6b7280;
-  }
-}
-
-// Controls Bar
-.controls-bar {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 2rem;
-  flex-wrap: wrap;
-}
-
-.search-section {
-  flex: 1;
-  display: flex;
-  gap: 0.75rem;
-  min-width: 300px;
-}
-
-.search-input-wrapper {
-  position: relative;
-  flex: 1;
-}
-
-.search-icon {
-  position: absolute;
-  left: 1rem;
-  top: 50%;
-  transform: translateY(-50%);
-  color: #9ca3af;
-  pointer-events: none;
-}
-
-.search-input {
-  width: 100%;
-  padding: 0.75rem 1rem 0.75rem 3rem;
-  border: 1px solid #e5e7eb;
-  border-radius: 0.5rem;
-  font-size: 1rem;
-  transition: all 0.2s;
-
-  &:focus {
-    outline: none;
-    border-color: #3b82f6;
-    box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
-  }
-
-  &::placeholder {
-    color: #9ca3af;
-  }
-}
-
-.search-button {
-  padding: 0.75rem 1.5rem;
-  background: #3b82f6;
-  color: white;
-  border: none;
-  border-radius: 0.5rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    background: #2563eb;
-  }
-
-  &:active {
-    transform: translateY(1px);
-  }
-}
-
-.create-workflow-btn {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1.5rem;
-  background: #10b981;
-  color: white;
-  border: none;
-  border-radius: 0.5rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    background: #059669;
-  }
-
-  svg {
-    width: 20px;
-    height: 20px;
-  }
-}
-
-// Tags Filter
-.tags-filter {
-  background: white;
-  padding: 1.5rem;
-  border-radius: 0.75rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  margin-bottom: 2rem;
-}
-
-.filter-label {
-  font-weight: 500;
-  color: #374151;
-  margin-right: 1rem;
-}
-
-.tag-list {
-  display: inline-flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.tag-chip {
-  padding: 0.375rem 0.875rem;
-  background: #f3f4f6;
-  color: #4b5563;
-  border: 1px solid transparent;
-  border-radius: 9999px;
-  font-size: 0.875rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    background: #e5e7eb;
-  }
-
-  &.active {
-    background: #dbeafe;
-    color: #1e40af;
-    border-color: #93c5fd;
-  }
-}
-
-// Loading State
-.loading-state {
-  text-align: center;
-  padding: 4rem;
-
-  .spinner {
-    width: 48px;
-    height: 48px;
-    border: 3px solid #f3f4f6;
-    border-top-color: #3b82f6;
-    border-radius: 50%;
-    margin: 0 auto 1rem;
-    animation: spin 1s linear infinite;
-  }
-
-  p {
-    color: #6b7280;
-    font-size: 1.125rem;
-  }
-}
-
-@keyframes spin {
-  to { transform: rotate(360deg); }
-}
-
-// Workflows Grid
-.workflows-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
-  gap: 1.5rem;
-  margin-bottom: 3rem;
-}
-
-.workflow-card {
-  background: white;
-  border-radius: 0.75rem;
-  overflow: hidden;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s;
-  cursor: pointer;
-
-  &:hover {
-    transform: translateY(-4px);
-    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.15);
-  }
-
-  &.enrolled {
-    border: 2px solid #10b981;
-  }
-}
-
-.card-thumbnail {
-  position: relative;
-  height: 160px;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  overflow: hidden;
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  .placeholder-thumbnail {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    svg {
-      width: 64px;
-      height: 64px;
-      color: rgba(255, 255, 255, 0.8);
-    }
-  }
-
-  .enrolled-badge {
-    position: absolute;
-    top: 0.75rem;
-    right: 0.75rem;
-    background: #10b981;
-    color: white;
-    padding: 0.375rem 0.75rem;
-    border-radius: 0.375rem;
-    font-size: 0.75rem;
-    font-weight: 500;
-    display: flex;
-    align-items: center;
-    gap: 0.375rem;
-  }
-}
-
-.card-content {
-  padding: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  height: calc(100% - 160px);
-}
-
-.card-title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #1f2937;
-  margin-bottom: 0.5rem;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.card-description {
-  color: #6b7280;
-  font-size: 0.875rem;
-  line-height: 1.5;
-  margin-bottom: 1rem;
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  flex: 1;
-}
-
-.card-tags {
-  display: flex;
-  gap: 0.375rem;
-  margin-bottom: 1rem;
-  flex-wrap: wrap;
-
-  .tag {
-    padding: 0.25rem 0.5rem;
-    background: #e0e7ff;
-    color: #4338ca;
-    border-radius: 0.25rem;
-    font-size: 0.75rem;
-    font-weight: 500;
-  }
-}
-
-.card-meta {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 1rem;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.meta-item {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  color: #6b7280;
-  font-size: 0.875rem;
-
-  svg {
-    width: 16px;
-    height: 16px;
-  }
-}
-
-.card-footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: auto;
-}
-
-.creator-info {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.creator-avatar {
-  width: 32px;
-  height: 32px;
-  background: linear-gradient(135deg, #3b82f6, #8b5cf6);
-  color: white;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: 600;
-  font-size: 0.875rem;
-}
-
-.creator-name {
-  color: #4b5563;
-  font-size: 0.875rem;
-}
-
-.action-button {
-  padding: 0.5rem 1rem;
-  background: #3b82f6;
-  color: white;
-  border: none;
-  border-radius: 0.375rem;
-  font-weight: 500;
-  font-size: 0.875rem;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    background: #2563eb;
-  }
-
-  &.enrolled {
-    background: #8b5cf6;
-
-    &:hover {
-      background: #7c3aed;
-    }
-  }
-}
-
-// Empty State
-.empty-state {
-  text-align: center;
-  padding: 4rem;
-
-  svg {
-    color: #d1d5db;
-    margin: 0 auto 1.5rem;
-  }
-
-  h3 {
-    font-size: 1.5rem;
-    font-weight: 600;
-    color: #374151;
-    margin-bottom: 0.5rem;
-  }
-
-  p {
-    color: #6b7280;
-    margin-bottom: 2rem;
-  }
-}
-
-// Pagination
-.pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-  margin-top: 3rem;
-}
-
-.page-btn {
-  padding: 0.5rem;
-  background: white;
-  border: 1px solid #e5e7eb;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  transition: all 0.2s;
-  display: flex;
-  align-items: center;
-
-  &:hover:not(:disabled) {
-    background: #f3f4f6;
-    border-color: #d1d5db;
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-
-  svg {
-    width: 20px;
-    height: 20px;
-  }
-}
-
-.page-info {
-  color: #6b7280;
-  font-weight: 500;
-}
-
-// Responsive
-@media (max-width: 768px) {
-  .browse-workflows {
-    padding: 1rem;
-  }
-
-  .page-header {
-    h1 {
-      font-size: 2rem;
-    }
-  }
-
-  .controls-bar {
-    flex-direction: column;
-  }
-
-  .search-section {
-    width: 100%;
-  }
-
-  .create-workflow-btn {
-    width: 100%;
-    justify-content: center;
-  }
-
-  .workflows-grid {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .tag-list {
-    display: flex;
-    overflow-x: auto;
-    padding-bottom: 0.5rem;
-    -webkit-overflow-scrolling: touch;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-}
