+++ frontend/src/app/shared/components/navigation/navigation.component.html
@@ -94,7 +94,7 @@
             Flashcards
           </a>
           <a
-            routerLink="/workflows"
+            routerLink="/courses"
             class="dropdown-item"
             routerLinkActive="active"
             (click)="closeMenu()"
@@ -112,7 +112,7 @@
                 d="M12 6v6m0 0v6m0-6h6m-6 0H6"
               />
             </svg>
-            Workflows
+            Courses
           </a>
         </div>
       </div>
